requirements.txt
accelerate==1.7.0
ace-tools==0.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
datasets==3.6.0
decorator==5.2.1
dill==0.3.8
distro==1.9.0
einops==0.8.1
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.0
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
hyperframe==6.1.0
idna==3.10
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.0
markupsafe==3.0.2
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
openai==1.79.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
portalocker==2.10.1
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.4
pydantic-core==2.33.2
pyte==0.8.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
qdrant-client==1.14.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
thefuck==3.32
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.0
